Winona computer vision (#61)
* Added thumbnail example

* Added thumbnail example (#59)

Renamed the recognize text example to Read Batch File (both local and remote), so not to be confused with an API call that is name recognize_text(). No code was changed. There is also the Recognize text for OCR that was left as is, since it does not use the read_batch_file() API call.

* Added Batch Read File example

Updated the comments/instructions to include it.

* Reformatted comments and code printouts

* Added Generate Thumbnail examples