Update quickstart to use V2 API only
The V2 API lets the user query a KB without using a separate runtime endpoint. It requires the user select "Managed" when creating the QnA Maker resource. We already have a quickstart for the V1 API, so this quickstart now shows the use of the V2 API only. This brings it into line with the Java and Python SDK quickstarts.