Adding Image URL to Identify Duplicate Images
In line 80 , I am proposing to have it as following - to add the file name which helps identify which images are seen duplicate:  
print(image.source_url,"Image status: ", image.status) 

This helps as the portal automatically removes the duplicates without identifying them and for the Custom SDK code, the program exits whenever there are duplicates. This proposed change helps the user remove duplicate images to ensure the smooth running of the rest of the code.